Clean up model.material.__init__
- Move build function to build.py
- Move CLI commands to separate submodule
- Remove model.material.build.apply_spec()
- Use model.build.apply_spec()
