+++ message_ix_models/cli.py
@@ -163,7 +163,7 @@ submodules = [
     "message_ix_models.model.water.cli",
     "message_ix_models.project.ssp",
     "message_ix_models.report.cli",
-    "message_ix_models.model.material",
+    "message_ix_models.model.material.cli",
     "message_ix_models.testing.cli",
     "message_ix_models.util.pooch",
 ]
